refactor: Extract expressions, patterns and constraints from parser
